requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.4
python-socketio==5.9.0
python-engineio==4.7.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
prometheus-flask-exporter==0.22.4
gunicorn==21.2.0
Werkzeug==2.3.7
click==8.1.7
colorama==0.4.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
SQLAlchemy==2.0.21